Section 1 - Project Setup
Downloaded and installed the version of Godot that is compatible with C#. Created a new project called DungeonMasterC# with support for version control (obviously). Imported all the assets that will be needed for this project so this first commit will be relatively large. Next commit will focus on setting up Player Movement.