Section 2 - Player Movement
Added Inputs in Godot for moving right, left, forward and backward. Created three animations for the Player including Move, Dash and Idle. Coded animations to play during moving and idling of the Player. Created a GameConstants.cs file to hold strings that will be used throughout the game. Added additional code to flip the Player when moving left and right. Finally, optimized the sprites of the animations for a smoother look. The next commit will focus on creating a State Machine to handle more complex movement of Characters in the game.