Section 3 - State Machine
Implemented a State Machine on the Player node and added three states, Idle, Move & Dash. Created an abstract state for the Player (PlayerState.cs) that all other Player States inherit from to control transitions between states with a minimal amount of code. In the next commit we will create a level that the Player can move around in.