Section 4 - Designing A Level
Using a grid map, designed a level for the Player to move around in. Attached a camera directly to the Player so the camera would follow during the game. Added a collision capsule to the Player so he doesn't walk through walls or stairs. Finally added some invisible walls to the level so the Player doesn't get trapped in certain spaces. Next commit will focus on creating Enemies for the Player to fight.