Section 5 - Adding Enemies
Created an Enemy scene complete with State Machine. The Enemy State Machine has six states: Idle, Move, Patrol, Return, Chase and Attack. Added a Navigation Mesh to the Environment scene for the Enemy to walk on and detect patrol points. Created collision spheres for detection of the Player for chasing and attacking. Did a lot of refactoring of all Character State Machines for better use of Inheritance.  Next commit will focus on the combat system.