Section 6 - Combat System
Implemented a combat system where the Player and Enemy can cause damage to each other. Started by adding animations for combo attacks for the Player. Added Area3D boxes for taking and causing damage to both characters. Toggled collision shapes of the make damage areas on and off depending on whether the character is attacking. Added Stats for the characters to include Health and Strength. Added Death states and Death animations for both characters. Characters will play the Death animation and be removed from the game when Health stat are equal or less than zero.